finished with plant suggestion, ready to pull from main
